+++ plugins/authenticator/guest-js/index.ts
@@ -17,7 +17,7 @@ export class Authenticator {
     challenge: string,
     application: string,
     registerData: string,
-    clientData: string
+    clientData: string,
   ): Promise<string> {
     return await invoke("plugin:authenticator|verify_registration", {
       challenge,
@@ -30,7 +30,7 @@ export class Authenticator {
   async sign(
     challenge: string,
     application: string,
-    keyHandle: string
+    keyHandle: string,
   ): Promise<string> {
     return await invoke("plugin:authenticator|sign", {
       timeout: 10000,
@@ -46,7 +46,7 @@ export class Authenticator {
     signData: string,
     clientData: string,
     keyHandle: string,
-    pubkey: string
+    pubkey: string,
   ): Promise<number> {
     return await invoke("plugin:authenticator|verify_signature", {
       challenge,