+++ plugins/fs-watch/guest-js/index.ts
@@ -45,7 +45,7 @@ async function unwatch(id: number): Promise<void> {
 export async function watch(
   paths: string | string[],
   cb: (event: DebouncedEvent) => void,
-  options: DebouncedWatchOptions = {}
+  options: DebouncedWatchOptions = {},
 ): Promise<UnlistenFn> {
   const opts = {
     recursive: false,
@@ -71,7 +71,7 @@ export async function watch(
     `watcher://debounced-event/${id}`,
     (event) => {
       cb(event.payload);
-    }
+    },
   );
 
   return () => {
@@ -83,7 +83,7 @@ export async function watch(
 export async function watchImmediate(
   paths: string | string[],
   cb: (event: RawEvent) => void,
-  options: WatchOptions = {}
+  options: WatchOptions = {},
 ): Promise<UnlistenFn> {
   const opts = {
     recursive: false,
@@ -109,7 +109,7 @@ export async function watchImmediate(
     `watcher://raw-event/${id}`,
     (event) => {
       cb(event.payload);
-    }
+    },
   );
 
   return () => {